+++ packages/@tinacms/git-client/src/index.ts
@@ -17,4 +17,6 @@ limitations under the License.
 */
 
 export * from './git-client'
+export * from './git-file'
+export * from './use-git-file'
 export * from './git-media-store'