+++ packages/@tinacms/git-client/src/use-git-file.ts
@@ -15,5 +15,21 @@ See the License for the specific language governing permissions and
 limitations under the License.
 
 */
+import { useMemo } from 'react'
+import { useCMS } from '@tinacms/react-core'
+import { GitFile } from './git-file'
 
-module.exports = require('../../typedoc.js')(require('./package.json'))
+export function useGitFile(
+  relativePath: string,
+  format: (file: any) => string,
+  parse: (content: string) => any
+) {
+  const cms = useCMS()
+
+  return useMemo(() => new GitFile(cms, relativePath, format, parse), [
+    cms,
+    relativePath,
+    format,
+    parse,
+  ])
+}