+++ packages/@tinacms/react-toolbar/src/components/Toolbar.tsx
@@ -93,6 +93,17 @@ export const Toolbar = () => {
     }
   }
 
+  /**
+   * TODO: decide best button defaults
+   * Should match defaults in form.ts
+   */
+  // @ts-ignore
+  const buttons = cms.toolbar?.buttons ||
+    form?.buttons || {
+      save: 'Save',
+      reset: 'Reset',
+    }
+
   //const reset = form && (form.reset || (() => form.finalForm.reset()))
   const submit = form && form.submit
   const disabled = !form
@@ -141,7 +152,7 @@ export const Toolbar = () => {
               onClick={reset}
             >
               <ResetIcon />
-              <DesktopLabel> Discard</DesktopLabel>
+              <DesktopLabel>{buttons.reset}</DesktopLabel>
             </ToolbarButton>
             <SaveButton
               primary
@@ -153,7 +164,7 @@ export const Toolbar = () => {
               {submitting && <LoadingDots />}
               {!submitting && (
                 <>
-                  Save <DesktopLabel>&nbsp;Page</DesktopLabel>
+                  <DesktopLabel>{buttons.save}</DesktopLabel>
                 </>
               )}
             </SaveButton>