+++ packages/@tinacms/react-core/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.4-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/react-core@0.3.3...@tinacms/react-core@0.3.4-alpha.0) (2020-07-15)
+
+**Note:** Version bump only for package @tinacms/react-core
+
+
+
+
+
+## [0.3.3](https://github.com/tinacms/tinacms/compare/@tinacms/react-core@0.3.3-alpha.0...@tinacms/react-core@0.3.3) (2020-07-07)
+
+**Note:** Version bump only for package @tinacms/react-core
+
+
+
+
+
+## [0.3.3-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/react-core@0.3.2...@tinacms/react-core@0.3.3-alpha.0) (2020-07-04)
+
+**Note:** Version bump only for package @tinacms/react-core
+
+
+
+
+
 ## [0.3.2](https://github.com/tinacms/tinacms/compare/@tinacms/react-core@0.3.1...@tinacms/react-core@0.3.2) (2020-06-29)
 
 **Note:** Version bump only for package @tinacms/react-core