+++ packages/gatsby-tinacms-remark/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.7-alpha.0](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-remark@0.8.6...gatsby-tinacms-remark@0.8.7-alpha.0) (2020-07-15)
+
+**Note:** Version bump only for package gatsby-tinacms-remark
+
+
+
+
+
+## [0.8.6](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-remark@0.8.6-alpha.0...gatsby-tinacms-remark@0.8.6) (2020-07-07)
+
+**Note:** Version bump only for package gatsby-tinacms-remark
+
+
+
+
+
+## [0.8.6-alpha.0](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-remark@0.8.5...gatsby-tinacms-remark@0.8.6-alpha.0) (2020-07-04)
+
+**Note:** Version bump only for package gatsby-tinacms-remark
+
+
+
+
+
 ## [0.8.5](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-remark@0.8.5-alpha.0...gatsby-tinacms-remark@0.8.5) (2020-06-29)
 
 **Note:** Version bump only for package gatsby-tinacms-remark