+++ packages/next-tinacms-markdown/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.24-alpha.0](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.23...next-tinacms-markdown@0.2.24-alpha.0) (2020-07-15)
+
+**Note:** Version bump only for package next-tinacms-markdown
+
+
+
+
+
+## [0.2.23](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.23-alpha.0...next-tinacms-markdown@0.2.23) (2020-07-07)
+
+**Note:** Version bump only for package next-tinacms-markdown
+
+
+
+
+
+## [0.2.23-alpha.0](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.22...next-tinacms-markdown@0.2.23-alpha.0) (2020-07-04)
+
+**Note:** Version bump only for package next-tinacms-markdown
+
+
+
+
+
 ## [0.2.22](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.22-alpha.0...next-tinacms-markdown@0.2.22) (2020-06-29)
 
 **Note:** Version bump only for package next-tinacms-markdown